pyproject.toml
[project]
name = "ankiconnect-server"
version = "0.1.0"
description = "Lightweight AnkiConnect Server"
authors = [{ name = "User", email = "user@example.com" }]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "anki>=25.2.6,<26",
    "flask>=2.3.3,<3",
    "flask-cors>=4.0.0,<5",
    "waitress>=2.1.2,<3",
    "requests>=2.31.0,<3",
    "jsonschema>=4.19.2,<5",
    "python-dotenv>=1.1.1,<2",
    "coverage>=7.9.2",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0,<8",
    "ruff-lsp>=0.0.62",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["E501"]  # Let formatter handle long lines

[tool.ruff.format]
